+++ src/main/java/com/example/goENC/dto/survey/reviseSurvey/RequestReviseSurveyDto.java
@@ -11,12 +11,12 @@ import java.util.List;
 @NoArgsConstructor
 public class RequestReviseSurveyDto {
     private User userId = null;
-    private Integer surveyId = null;
+    private Long surveyId = null;
     private String surveyTitle = null;
     private String surveyContent = null;
     private List<RequestQuestionDto> questionCardList = null;
 
-    public RequestReviseSurveyDto(Long userId, Integer surveyId, String surveyTitle, String surveyContent, List<RequestQuestionDto> questionCardList) {
+    public RequestReviseSurveyDto(Long userId, Long surveyId, String surveyTitle, String surveyContent, List<RequestQuestionDto> questionCardList) {
         this.userId = new User(userId);
         this.surveyId = surveyId;
         this.surveyTitle = surveyTitle;
